circuits: Extract common test setup and use production CRS
Move shared test infrastructure (TestingParameters, helper functions)
into a common module. Load the production CRS from cache or download
it, instead of generating a test CRS, to match the deployed system.
